NeMo tests refactor - merge into master (#454)
* work on unit tests - mocking up the i/o read/wrote


* work on unit tests - mocking up the i/o read/wrote


* a file write mockup


* doc fix of configuration tutorials


* reorganization of directories in tests


* moved some tests to integration, fixed dirs in tests/unit/core/test_weight_share.py


* Cleaning up unneccesary inits, adding markers, pytest -m


* markers for different tests, docs marked


* first approach to adding run-time flags to pytest


* Passing cpu flag to neural factory fixture working


* Neural module unit test refactor - middle of core unit


* neural types unit tests


* Added fixture enabling to skip some of the tests considering selected device, refactored trainers and export


* finished refactoring core tests


* Tests using separate trainers - not accessible from NeuralFactory


* Refactored TTS and ASR tests, split into acceptance and unit


* Cleanup, NLP tests refactored/reorganized


* polishing, added two missing markers


* tacotron training disabled on CPU


* style fix


* Jenkins updated - reflecting new organization of tests


* NeMo Test Reorganization (#430)

* reorganization of directories in tests


* moved some tests to integration, fixed dirs in tests/unit/core/test_weight_share.py


* Cleaning up unneccesary inits, adding markers, pytest -m


* markers for different tests, docs marked


* first approach to adding run-time flags to pytest


* Passing cpu flag to neural factory fixture working


* Neural module unit test refactor - middle of core unit


* neural types unit tests


* Added fixture enabling to skip some of the tests considering selected device, refactored trainers and export


* finished refactoring core tests


* Tests using separate trainers - not accessible from NeuralFactory


* Refactored TTS and ASR tests, split into acceptance and unit


* Cleanup, NLP tests refactored/reorganized


* polishing, added two missing markers


* tacotron training disabled on CPU


* style fix


* Jenkins updated - reflecting new organization of tests


* Introduced  mark, reclassified all acceptance to unclassified


* unclassified jenkins fix


* touch


* touch


* signoff


* after git reset HEAD~5
