+++ setup.cfg
@@ -1,9 +1,31 @@
+# =============================================================================
+# Copyright (c) 2020, NVIDIA CORPORATION.  All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# =============================================================================
+
 [aliases]
 test=pytest
 
 [tool:pytest]
-addopts = --verbose --pyargs
-testpaths = tests
+addopts = --verbose --pyargs 
+markers =
+    unit: marks unit test, i.e. testing a single, well isolated functionality (deselect with '-m "not unit"')
+    integration: marks test checking the elements when integrated into subsystems (deselect with '-m "not integration"')
+    system: marks test working at the highest integration level. (deselect with '-m "not system"')
+    acceptance: marks test checking whether the developed product/model passes the user defined acceptance criteria (deselect with '-m "not acceptance"')
+    docs: mark tests related to documentation (deselect with '-m "not docs"')
+    unclassified: marks tests that are unclassified and need to be fixed/rewritten (deselect with '-m "not unclassified"')
 
 [isort]
 known_localfolder = nemo,tests
