+++ tests/__init__.py
@@ -1,6 +1,7 @@
 # ! /usr/bin/python
 # -*- coding: utf-8 -*-
 
+# =============================================================================
 # Copyright 2020 NVIDIA. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
