+++ tests/system/test_infer.py
@@ -16,13 +16,15 @@
 # limitations under the License.
 # =============================================================================
 
+from unittest import TestCase
+
+import pytest
 import torch
 
 import nemo
 from nemo.backends.pytorch.nm import NonTrainableNM
-from nemo.core.neural_types import *
+from nemo.core.neural_types import AxisKind, AxisType, ChannelType, NeuralType
 from nemo.utils.decorators import add_port_docs
-from tests.common_setup import NeMoUnitTest
 
 
 class AddsTen(NonTrainableNM):
@@ -63,12 +65,15 @@ class SubtractsTen(NonTrainableNM):
         return mod_in - 10
 
 
-class TestInfer(NeMoUnitTest):
-    def test_infer_caching(self):
-        neural_factory = nemo.core.neural_factory.NeuralModuleFactory(
-            backend=nemo.core.Backend.PyTorch, create_tb_writer=False
-        )
+@pytest.mark.usefixtures("neural_factory")
+class TestInfer(TestCase):
+    def setUp(self) -> None:
+        """ Re-instantiates Neural Factory for every test. """
+        # Re-initialize the default Neural Factory - on the indicated device.
+        self.nf = nemo.core.NeuralModuleFactory(placement=self.nf.placement)
 
+    @pytest.mark.system
+    def test_infer_caching(self):
         data_source = nemo.backends.pytorch.common.ZerosDataLayer(
             size=1,
             dtype=torch.FloatTensor,
@@ -85,18 +90,16 @@ class TestInfer(NeMoUnitTest):
         twenty_tensor = addten(mod_in=ten_tensor)
         thirty_tensor = addten(mod_in=twenty_tensor)
 
-        evaluated_tensors = neural_factory.infer(tensors=[twenty_tensor, thirty_tensor], verbose=False, cache=True)
+        evaluated_tensors = self.nf.infer(tensors=[twenty_tensor, thirty_tensor], verbose=False, cache=True)
         self.assertEqual(evaluated_tensors[0][0].squeeze().data, 20)
         self.assertEqual(evaluated_tensors[1][0].squeeze().data, 30)
 
         new_ten_tensor = minusten(mod_in=twenty_tensor)
-        evaluated_tensors = neural_factory.infer(tensors=[new_ten_tensor], verbose=False, use_cache=True)
+        evaluated_tensors = self.nf.infer(tensors=[new_ten_tensor], verbose=False, use_cache=True)
         self.assertEqual(evaluated_tensors[0][0].squeeze().data, 10)
 
+    @pytest.mark.system
     def test_infer_errors(self):
-        neural_factory = nemo.core.neural_factory.NeuralModuleFactory(
-            backend=nemo.core.Backend.PyTorch, create_tb_writer=False
-        )
 
         data_source = nemo.backends.pytorch.common.ZerosDataLayer(
             size=1,
@@ -115,21 +118,19 @@ class TestInfer(NeMoUnitTest):
         thirty_tensor = addten(mod_in=twenty_tensor)
 
         with self.assertRaisesRegex(ValueError, "use_cache was set, but cache was empty"):
-            evaluated_tensors = neural_factory.infer(
-                tensors=[twenty_tensor, thirty_tensor], verbose=False, use_cache=True
-            )
+            evaluated_tensors = self.nf.infer(tensors=[twenty_tensor, thirty_tensor], verbose=False, use_cache=True)
 
         new_ten_tensor = minusten(mod_in=twenty_tensor)
-        evaluated_tensors = neural_factory.infer(tensors=[new_ten_tensor], verbose=False, cache=True)
+        evaluated_tensors = self.nf.infer(tensors=[new_ten_tensor], verbose=False, cache=True)
 
         with self.assertRaisesRegex(ValueError, "cache was set but was not empty"):
-            evaluated_tensors = neural_factory.infer(tensors=[twenty_tensor, thirty_tensor], verbose=False, cache=True)
+            evaluated_tensors = self.nf.infer(tensors=[twenty_tensor, thirty_tensor], verbose=False, cache=True)
 
-        neural_factory.clear_cache()
-        evaluated_tensors = neural_factory.infer(tensors=[new_ten_tensor], verbose=False, cache=True)
+        self.nf.clear_cache()
+        evaluated_tensors = self.nf.infer(tensors=[new_ten_tensor], verbose=False, cache=True)
 
         with self.assertRaisesRegex(ValueError, "cache and use_cache were both set."):
-            evaluated_tensors = neural_factory.infer(
+            evaluated_tensors = self.nf.infer(
                 tensors=[twenty_tensor, thirty_tensor], verbose=False, cache=True, use_cache=True
             )
         self.assertEqual(evaluated_tensors[0][0].squeeze().data, 10)
